fix inlinetest package organization
